Revert "Split main thread webview serializer code into own file"
This reverts commit 04de4655246c9760fd69dcd4dd195f6b2feb74b2.
